+++ src/main/kotlin/com/possible_triangle/bigsip/alcohol/IAlcoholLevel.kt
@@ -2,7 +2,7 @@ package com.possible_triangle.bigsip.alcohol
 
 interface IAlcoholLevel {
 
-    var current: Int
+    var current: Float
 
     var persistent: Float
